refactor code generator's chat structure (#78)
Moving original chat content from "executor" to "user". Now, the user is
responsible to report a feedback to the CG. One consequence is that we
need to add an extra post from the user if the end of the conversation
is "ci->user". Because otherwise, there will be no execution result.

This can only happen in examples or during chat history summarization. 

misc:
- changed cg's prompt file name
- fixed typos in prompts
- fixed old uts and add new uts